+++ node-wasm/src/worker.rs
@@ -1,12 +1,14 @@
 use std::error::Error;
 use std::fmt::{self, Debug, Display};
+use std::marker::PhantomData;
 
 use futures::future::{BoxFuture, FutureExt, TryFutureExt};
 use libp2p::Multiaddr;
+use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use serde_wasm_bindgen::{from_value, to_value};
 use tokio::sync::{mpsc, oneshot};
-use tracing::{info, trace, warn};
+use tracing::{error, info, trace, warn};
 use wasm_bindgen::prelude::*;
 use wasm_bindgen_futures::spawn_local;
 use web_sys::{MessageEvent, MessagePort, SharedWorker};
@@ -19,7 +21,7 @@ use lumina_node::store::{IndexedDbStore, SamplingMetadata, Store};
 use lumina_node::syncer::SyncingInfo;
 
 use crate::node::WasmNodeConfig;
-use crate::utils::{CommandResponseChannel, NodeCommandResponse, NodeCommandType, WorkerSelf};
+use crate::utils::WorkerSelf;
 use crate::wrapper::libp2p::NetworkInfoSnapshot;
 
 type Result<T, E = WorkerError> = std::result::Result<T, E>;
@@ -39,6 +41,106 @@ impl From<NodeError> for WorkerError {
         WorkerError(error.to_string())
     }
 }
+
+pub type CommandResponseChannel<T> = oneshot::Sender<<T as NodeCommandType>::Output>;
+#[derive(Serialize, Deserialize, Debug)]
+pub struct NodeCommandResponse<T>(pub T::Output)
+where
+    T: NodeCommandType,
+    T::Output: Debug + Serialize;
+
+pub trait NodeCommandType: Debug + Into<NodeCommand> {
+    type Output;
+}
+
+pub struct BChannel<IN, OUT> {
+    _onmessage: Closure<dyn Fn(MessageEvent)>,
+    _forwarding_task: (),
+    channel: MessagePort,
+    response_channels: mpsc::Sender<oneshot::Sender<OUT>>,
+    send_type: PhantomData<IN>,
+}
+
+impl<IN, OUT> BChannel<IN, OUT>
+where
+    IN: Serialize,
+    // XXX: send sync shouldn't be needed
+    OUT: DeserializeOwned + 'static + Send + Sync,
+{
+    pub fn new(channel: MessagePort) -> Self {
+        let (tx, mut message_rx) = mpsc::channel(64);
+
+        let near_tx = tx.clone();
+        let f = move |ev: MessageEvent| {
+            let local_tx = near_tx.clone();
+            spawn_local(async move {
+                let message_data = ev.data();
+
+                let data = from_value(message_data).unwrap();
+
+                local_tx.send(data).await.expect("send err");
+            })
+        };
+
+        let (response_tx, mut response_rx) = mpsc::channel(1);
+        let forwarding_task = spawn_local(async move {
+            loop {
+                let message = message_rx.recv().await.unwrap();
+                let response_channel: oneshot::Sender<OUT> = response_rx.recv().await.unwrap();
+                response_channel
+                    .send(message)
+                    .ok()
+                    .expect("forwarding broke");
+            }
+        });
+
+        let onmessage = Closure::new(f);
+
+        channel.set_onmessage(Some(onmessage.as_ref().unchecked_ref()));
+        channel.start();
+        Self {
+            response_channels: response_tx,
+            _onmessage: onmessage,
+            _forwarding_task: forwarding_task,
+            channel,
+            send_type: PhantomData,
+        }
+    }
+
+    pub fn send<T: NodeCommandType>(
+        &self,
+        command: T,
+    ) -> BoxFuture<'static, Result<T::Output, tokio::sync::oneshot::error::RecvError>>
+    where
+        T::Output: Debug + Serialize,
+        NodeCommandResponse<T>: TryFrom<OUT>,
+        <NodeCommandResponse<T> as TryFrom<OUT>>::Error: Debug,
+    {
+        let message: NodeCommand = command.into();
+        self.send_enum(message);
+
+        let (tx, rx) = oneshot::channel();
+
+        self.response_channels.try_send(tx).expect("busy");
+
+        // unwrap "handles" invalid type
+        rx.map_ok(|r| {
+            tracing::info!("type = , expected = {}", std::any::type_name::<T>());
+            NodeCommandResponse::<T>::try_from(r)
+                .expect("invalid type")
+                .0
+        })
+        .boxed()
+    }
+
+    fn send_enum(&self, msg: NodeCommand) {
+        let v = to_value(&msg).unwrap();
+        self.channel.post_message(&v).expect("err post");
+    }
+}
+
+// TODO: cleanup JS objects on drop
+// impl Drop
 
 macro_rules! define_command_from_impl {
     ($common_name:ident, $command_name:ident) => {
@@ -299,7 +401,6 @@ impl NodeWorker {
             NodeCommandWithChannel::GetHeader((command, response)) => {
                 let header = match command.0 {
                     SingleHeaderQuery::Head => self.node.get_local_head_header().await.unwrap(),
-                    //SingleHeaderQuery::NetworkHead => self.node.get_network_head_header().unwrap(),
                     SingleHeaderQuery::ByHash(hash) => {
                         self.node.get_header_by_hash(&hash).await.ok().unwrap()
                     }
@@ -485,5 +586,5 @@ pub async fn run_worker(queued_connections: Vec<MessagePort>) {
         }
     }
 
-    warn!("EXIT EXIT EXIT");
+    error!("Worker exited, should not happen");
 }
